+++ src/main/java/personal_project/moment_talk/chat/controller/ChatController.java
@@ -4,15 +4,12 @@ import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 import personal_project.moment_talk.chat.dto.CreateGroupRequest;
 import personal_project.moment_talk.chat.service.DeepLTranslationService;
 import personal_project.moment_talk.common.redis.GroupChatParticipants;
-import personal_project.moment_talk.common.webSocket.GroupChatWebSocketHandler;
 
 import java.util.List;
 import java.util.Map;
@@ -24,7 +21,6 @@ public class ChatController {
 
     private final DeepLTranslationService deepLTranslationService;
     private final GroupChatParticipants groupChatParticipants;
-    private final GroupChatWebSocketHandler groupChatWebSocketHandler;
     private final RedisTemplate redisTemplate;
 
     @GetMapping("/1-to-1-chat")
@@ -48,19 +44,15 @@ public class ChatController {
             return ResponseEntity.badRequest().body(Map.of("message", "Invalid room ID!"));
         }
 
-        // 그룹 채팅방 참가 로직
-        groupChatWebSocketHandler.handleJoinRoom(roomId, httpSessionId);
-
+        groupChatParticipants.joinGroupChatRoom(roomId, httpSessionId);
         return ResponseEntity.ok(Map.of("message", "User joined the room", "roomId", roomId));
     }
 
     @PostMapping("/group-chat/leave")
     public ResponseEntity<String> leaveGroupChat(HttpSession httpSession) {
-        String roomId = (String) redisTemplate.opsForHash().get("chat:session_to_room", httpSession.getId());
-        if (roomId == null || roomId.isEmpty()) {
+        if (!groupChatParticipants.leaveGroupChatRoom(httpSession.getId())) {
             return ResponseEntity.ok("단체 채팅방에 들어가 있지 않습니다");
         }
-        groupChatWebSocketHandler.handleLeaveRoom(roomId, httpSession.getId());
         return ResponseEntity.ok("단체 채팅방에서 나갔습니다");
     }
 
@@ -76,9 +68,7 @@ public class ChatController {
         String roomName = request.name();
         String httpSessionId = httpSession.getId();
 
-        log.info("Creating room: {} by session: {}", roomName, httpSessionId);
-
-        groupChatWebSocketHandler.handleCreateRoom(roomName, httpSessionId);
+        groupChatParticipants.createGroupChatRoom(roomName, httpSessionId);
 
         String roomId = (String) redisTemplate.opsForHash().get("chat:rooms", roomName);
         log.info("Created room ID: {}", roomId);
