+++ src/main/java/personal_project/moment_talk/common/redis/GroupChatParticipants.java
@@ -5,7 +5,6 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
 @Component
@@ -17,6 +16,7 @@ public class GroupChatParticipants {
     private static final String PARTICIPANTS_KEY_PREFIX = "chat:room:";
     private static final String PRIVATE_ROOM_ID = "chat:session_to_room";
 
+    // 그룹 채팅 생성
     public void createGroupChatRoom(String roomName, String httpSessionId) {
         String roomId = UUID.randomUUID().toString(); // 방 ID 생성
         redisTemplate.opsForHash().put(ROOM_KEY, roomName, roomId); // Redis에 저장
@@ -29,17 +29,21 @@ public class GroupChatParticipants {
 
         List<String> participants = new ArrayList<>();
         participants.add(httpSessionId);
-//        chatParticipants.put(roomId, participants);
 
         redisTemplate.opsForHash().put(PRIVATE_ROOM_ID, httpSessionId, roomId);
     }
 
-    public void deleteGroupChatRoom(String roomId) {
-        redisTemplate.opsForHash().delete(ROOM_KEY, roomId);
-        redisTemplate.delete(PARTICIPANTS_KEY_PREFIX + roomId + ":participants");
-    }
+    // 그룹 채팅에서 나가기
+    public boolean leaveGroupChatRoom(String httpSessionId) {
+        String roomId = (String) redisTemplate.opsForHash().get("chat:session_to_room", httpSessionId);
+        if (roomId == null || roomId.isEmpty()) {
+            return false;
+        }
+        removeParticipant(roomId, httpSessionId);
+        return true;
+        }
 
-    public void removeParticipant(String roomId, String httpSessionId) {
+    private void removeParticipant(String roomId, String httpSessionId) {
         String key = PARTICIPANTS_KEY_PREFIX + roomId + ":participants";
         redisTemplate.opsForSet().remove(key, httpSessionId);
         redisTemplate.opsForHash().delete(PRIVATE_ROOM_ID, httpSessionId);
@@ -59,7 +63,8 @@ public class GroupChatParticipants {
         }
     }
 
-    public void addParticipantToGroupChatRoom(String roomId, String httpSessionId) {
+    // 그룹 채팅 참가
+    public void joinGroupChatRoom(String roomId, String httpSessionId) {
         redisTemplate.opsForSet().add(PARTICIPANTS_KEY_PREFIX + roomId + ":participants", httpSessionId);
         redisTemplate.opsForHash().put(PRIVATE_ROOM_ID, httpSessionId, roomId);
     }
